adding nav bar to the top adding the logout button added theme changer
